did some progress on useState for currentBlock
