+++ src/components/tetris_bit/grid_resource/row.js
@@ -1,0 +1,11 @@
+export default function Row() {
+    const GRID_LENGTH = 6;
+
+    let row = [];
+    for (let i = 0; i < GRID_LENGTH; i++) {
+        row.push(
+            <div key={i} className="box"></div>
+        )
+    }
+    return row
+}
